Added new options and icons.
Added password option to the options screen, the options is used while
communicating to SSS. Added new set of icons, using the ones with
transparent background. Added a check for SSS configuration options
being set, displaying an error message if not. The popup.js file will
have to be reformatted in order to become readable.